feat(core): extract vec-store (KJC-TSK-0511)
PR4 of the KJC-TSK-0511 monorepo decoupling: lift the RAG vector store
(better-sqlite3 + sqlite-vec wrapper) into @karajan/core so the hu-board
workspace can read rag.db without reaching into src/rag/.

The original src/rag/vec-store.js becomes a 17-line shim re-exporting
the 12 named symbols. hu-board's two callers (routes/api.js, routes/rag.js)
import directly from @karajan/core/vec-store. better-sqlite3 and
sqlite-vec move into core/package.json as production deps.

New core smoke tests cover dbPath env precedence, schema init, round-trip
insert/search, deleteChunksBySource, projectSlug normalisation, content-hash
lookup scoped per-project, and the per-project commit watermark.
